Added tests for GamesService

+++ server/src/main/java/model/response/CreateGameResponse.java
@@ -1,4 +1,4 @@
 package model.response;
 
-public record CreateGameResponse(long gameID) {
+public record CreateGameResponse(int gameID) {
 }
